use git clone in cases git package exists;

+++ servers/monitoring/init-node-exporter.js
@@ -58,14 +58,22 @@ program.iterate('hosts', async (host) => {
         // Remove previous symlink
         await ssh.exec('rm -frv /opt/node_exporter') // temp
 
-        // Install (Some servers does not have git, so we rsync it instead)
-        // Starting local shell
-        let shell = await program.shell()
-        await shell.exec('rm -rf exporters') // delete locally
-        await program.chat.notify('Cloning exporters repo(Locally)')
-        await shell.exec('git clone git@gitlab.dopamine.bg:devops/monitoring/exporters.git')
-        await shell.exec(`rsync -vrltgoD --delete exporters root@${hostIP}:/opt/dopamine`)
-        await shell.exec('rm -rf exporters') // delete locally
+        if(await ssh.packageExists('git')) {
+            if(!await ssh.exists('/opt/dopamine/exporters/')) {
+                await ssh.exec('git clone git@gitlab.dopamine.bg:devops/monitoring/exporters.git /opt/dopamine/exporters')
+            }
+            await ssh.chdir('/opt/dopamine/exporters')
+            await ssh.exec('git pull')
+        } else {
+            // Install (Some servers does not have git, so we rsync it instead)
+            // Starting local shell
+            let shell = await program.shell()
+            await shell.exec('rm -rf exporters') // delete locally
+            await program.chat.notify('Cloning exporters repo(Locally)')
+            await shell.exec('git clone git@gitlab.dopamine.bg:devops/monitoring/exporters.git')
+            await shell.exec(`rsync -vrltgoD --delete exporters root@${hostIP}:/opt/dopamine`)
+            await shell.exec('rm -rf exporters') // delete locally
+        }
 
         await ssh.exec('chmod +x /opt/dopamine/exporters/node_exporter/node_exporter') // delete on server
